Implement variants, transformers and previews
This is another infra piece inside ActiveStorage. They allow to modify
image size and other transformations. Also activestorage offer preview
of other types of files like video or pdf. We want to bring all that
infra inside this package in the same way as a behavior so it can be
replaced.